fixed path error; ci test1

+++ examples/capiTester.f90
@@ -2,8 +2,8 @@ program name
 
     USE rosenna
     implicit none
-    REAL, DIMENSION(1,2) :: inputs
-    REAL, DIMENSION(    1, 3) :: output
+    REAL (c_double), DIMENSION(1,2) :: inputs
+    REAL (c_double), DIMENSION(    1, 3) :: output
 
     inputs = RESHAPE(    (/1.0, 1.0/),    (/1, 2/), order =     [2 , 1 ])
